% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{last_change}
\alias{last_change}
\title{Last Change}
\usage{
last_change(r, cores = 0)
}
\arguments{
\item{r}{SpatRaster}

\item{cores}{num of processor cores to be used}
}
\value{
SpatRaster
}
\description{
number of periods from last change
}
