% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{lu_change_classAB}
\alias{lu_change_classAB}
\title{lu_change_classAB}
\usage{
lu_change_classAB(r, classab, cores)
}
\arguments{
\item{r}{spatRaster or RasterStack with LU layers}

\item{classab}{vector 2 class id ex : c(4,6) or c('forest', 'pasture)}

\item{cores}{integer number of CPU cores}
}
\value{
statRaster
}
\description{
Determines by pixel how many times a class A has changed
to a class B
}
\examples{
library(MBprocess)
library(terra)
library(ggplot2)
lu <- system.file("extdata", 'lu_test.tif', package = "MBprocess")
mb <- terra::rast(lu)
rm1 <- MBprocess::lu_change_classAB(r = mb, classab = c(6,4), cores = 0)
terra::freq(rm1)
terra::plot(rm1)
}
