% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{lu_class_max_sequence_cont}
\alias{lu_class_max_sequence_cont}
\title{Class Maximum Sequence without LU change}
\usage{
lu_class_max_sequence_cont(r, cores = 0)
}
\arguments{
\item{r}{SpatRaster input rasters with one layer by year}

\item{cores}{integer number of cores to be used dot multitasking}
}
\value{
SpatRaster
}
\description{
Determines the Class id with  maximum sequence of years without land use change
}
\examples{
library(MBprocess)
library(terra)
lu <- system.file("extdata", 'lu_test.tif', package = "MBprocess")
mb <- terra::rast(lu)
mmax = lu_class_max_sequence_cont(r = mb, cores = 0)
fr = terra::freq(mmax)
print(fr)
}
