% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{lu_diversity}
\alias{lu_diversity}
\title{Diversity}
\usage{
lu_diversity(r, cores = 0)
}
\arguments{
\item{r}{SpatRaster input rasters with one layer by year}

\item{cores}{integer number of cores to be used to do multitasking}
}
\value{
SpatRaster
}
\description{
calculates the different number os class by cell/pixel
}
\examples{
#' library(MBprocess)
library(terra)
library(ggplot2)
lu <- system.file("extdata", 'lu_test.tif', package = "MBprocess")
mb <- terra::rast(lu)
plot(mb[[1]], main = names(mb)[1])
diversity = MBprocess::lu_diversity(r = mb, cores = 0)
plot(diversity)
writeRaster(diversity, filename = 'diversity.tif', overwrite = TRUE,
gdal = c("COMPRESS=LZW"), datatype = "INT1U")

df = terra::freq(diversity) |>
data.frame() |>
  subset(select = -layer) |>
  transform(freq_100 = count/sum(count) * 100)

ggplot(df, aes(x = as.factor(value), y = count)) +
  geom_col() + labs(x = 'diversity', y = 'number of pixels',
                    title = 'Diversity',
                    subtitle = 'by pixel',
                    caption = 'data by MapBiomas(2021)') +
  theme_bw()

ggplot(df, aes(x = as.factor(value), y = freq_100)) +
  geom_col() +
  geom_text(aes(label = round(freq_100, 1) ), nudge_y = 2, size = 3) +
  labs(x = 'number of changes', y = '\%',
       title = 'Number of different Land Use',
       subtitle = 'by pixel',
       caption = 'data by MapBiomas(2021)') +
  theme_bw()
}
