% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{lu_mode}
\alias{lu_mode}
\title{Land Use Mode}
\usage{
lu_mode(r, cores = 0)
}
\arguments{
\item{r}{SpatRaster}

\item{cores}{num of processors cores to be used}
}
\value{
SpatRaster with the id of modal class
}
\description{
Calculate Mode - the most frequent class in a stack or land use rasters
}
\examples{
library(MBprocess)
library(terra)
library(ggplot2)
lu <- system.file("extdata", 'lu_test.tif', package = "MBprocess")
mb <- terra::rast(lu)
plot(mb[[1]], main = names(mb)[1])
class_mode = lu_mode(r = mb, cores = 0)
plot(class_mode)
writeRaster(class_mode, filename = 'class_mode.tif', overwrite = TRUE,
            gdal = c("COMPRESS=LZW"), datatype = "INT1U")

df = terra::freq(class_mode) |>
  data.frame() |>
  subset(select = -layer) |>
  transform(freq_100 = count/sum(count) * 100)

ggplot(df, aes(x = as.factor(value), y = count)) +
  geom_col() + labs(x = 'LU Class', y = 'number of pixels',
                    title = 'Class Mode',
                    subtitle = 'by pixel',
                    caption = 'data by MapBiomas(2021)') +
  theme_bw()

ggplot(df, aes(x = as.factor(value), y = freq_100)) +
  geom_col() +
  geom_text(aes(label = round(freq_100, 1) ), nudge_y = 1, size = 3) +
  labs(x = 'LU Class', y = '\%',
       title = 'Class Mode',
       subtitle = 'by pixel',
       caption = 'data by MapBiomas(2021)') +
  theme_bw()
}
