% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(r, cores = 0)
}
\arguments{
\item{r}{SpatRaster}

\item{cores}{num of processor cores to be used}
}
\value{
SpatRaster
}
\description{
Calculate Mode - the more frequent class in a stack
}
