% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{lu_change_classAB}
\alias{lu_change_classAB}
\title{lu_change_classAB}
\usage{
lu_change_classAB(r, classa, classb, cores)
}
\arguments{
\item{classa}{lu class original}

\item{classb}{lu class changed}

\item{cores}{integer number of CPU cores}
}
\value{
statRaster
}
\description{
Determines by pixel how many times a class A has changed
to a class B
}
\examples{
library(MBprocess)
library(terra)
library(ggplot2)
lu <- system.file("extdata", 'lu_test.tif', package = "MBprocess")
mb <- terra::rast(lu)
rm1 <- terra::app(mb, vecIn,  b = c(4,6),  cores = 0)
terra::freq(rm1)
terra::plot(rm1)
}
