% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{lu_n_change}
\alias{lu_n_change}
\title{Number of LU changes}
\usage{
lu_n_change(r, cores = 0)
}
\arguments{
\item{r}{SpatRaster input rasters with one layer by year}

\item{cores}{integer number of cores to be used dot multitasking}
}
\value{
SpatRaster
}
\description{
Determines the number of times that land use change
}
\examples{
library(terra)
lu <- system.file("extdata", 'lu_test.tif', package = "MBprocess")
mb <- terra::rast(lu)
mchange = lu_n_change(r = mb, cores = 0)
fr = terra::freq(mchange)
print(fr)
}
